requirements.txt
plaidml-keras
plaidbench
tensorflow
matplotlib